requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Markdown==3.5.1
django-filter==23.5
Pillow==10.1.0 